src/data/usecases/account/db-load-account-by-token.ts
import { Decrypter, LoadAccountByTokenRepository } from '@/data/protocols'
import { LoadAccountByToken } from '@/domain/usecases'

export class DbLoadAccountByToken implements LoadAccountByToken {
  constructor (
    private readonly decrypter: Decrypter,
    private readonly loadAccountByTokenRepository: LoadAccountByTokenRepository,
  ) {}

  async load (accessToken: string, role?: string): Promise<LoadAccountByToken.Result> {
    const isTokenValid = await this.decrypter.decrypt(accessToken)
    if (isTokenValid) {
      const account = await this.loadAccountByTokenRepository.loadByToken(accessToken, role)
      if (account) {
        return account
      }
    }
    return null
  }
}
